third_party/WebKit/LayoutTests/fast/css/getFloatValueForUnit-expected.txt

Testing whether CSSPrimitiveValue.getFloatValue() converts units correctly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS getFloatValue(0, 'font-size', CSSPrimitiveValue.CSS_PT) is 10
PASS getFloatValue(0, 'font-size', CSSPrimitiveValue.CSS_MM) is 3.527778
PASS getFloatValue(0, 'font-size', CSSPrimitiveValue.CSS_CM) is 0.352778
PASS getFloatValue(0, 'font-size', CSSPrimitiveValue.CSS_IN) is 0.138889
PASS getFloatValue(0, 'font-size', CSSPrimitiveValue.CSS_PC) is 0.833333
PASS getFloatValue(0, 'font-size', CSSPrimitiveValue.CSS_PX) is 13.333333
PASS getFloatValue(0, 'font-size', CSSPrimitiveValue.CSS_NUMBER) is 13.333333
PASS getFloatValue(2, '-webkit-transform', CSSPrimitiveValue.CSS_DEG) is 90
PASS getFloatValue(2, '-webkit-transform', CSSPrimitiveValue.CSS_RAD) is 1.570796
PASS getFloatValue(2, '-webkit-transform', CSSPrimitiveValue.CSS_GRAD) is 100
PASS getFloatValue(2, '-webkit-transform', CSSPrimitiveValue.CSS_NUMBER) is 90
PASS getFloatValue(1, '-webkit-animation-delay', CSSPrimitiveValue.CSS_MS) is 200
PASS getFloatValue(1, '-webkit-animation-delay', CSSPrimitiveValue.CSS_S) is 0.2
PASS getFloatValue(1, '-webkit-animation-delay', CSSPrimitiveValue.CSS_NUMBER) is 200
PASS getFloatValue(0, 'font-size', CSSPrimitiveValue.CSS_PERCENTAGE) threw exception InvalidAccessError: Failed to execute 'getFloatValue' on 'CSSPrimitiveValue': Failed to obtain a double value..
PASS getFloatValue(0, 'font-size', CSSPrimitiveValue.CSS_MS) threw exception InvalidAccessError: Failed to execute 'getFloatValue' on 'CSSPrimitiveValue': Failed to obtain a double value..
PASS getFloatValue(0, 'font-size', CSSPrimitiveValue.CSS_HZ) threw exception InvalidAccessError: Failed to execute 'getFloatValue' on 'CSSPrimitiveValue': Failed to obtain a double value..
PASS getFloatValue(0, 'font-size', CSSPrimitiveValue.CSS_DEG) threw exception InvalidAccessError: Failed to execute 'getFloatValue' on 'CSSPrimitiveValue': Failed to obtain a double value..
PASS getFloatValue(0, 'font-size', CSSPrimitiveValue.CSS_RGBCOLOR) threw exception InvalidAccessError: Failed to execute 'getFloatValue' on 'CSSPrimitiveValue': Failed to obtain a double value..
PASS getFloatValue(2, '-webkit-transform', CSSPrimitiveValue.CSS_PERCENTAGE) threw exception InvalidAccessError: Failed to execute 'getFloatValue' on 'CSSPrimitiveValue': Failed to obtain a double value..
PASS getFloatValue(2, '-webkit-transform', CSSPrimitiveValue.CSS_PX) threw exception InvalidAccessError: Failed to execute 'getFloatValue' on 'CSSPrimitiveValue': Failed to obtain a double value..
PASS getFloatValue(2, '-webkit-transform', CSSPrimitiveValue.CSS_MS) threw exception InvalidAccessError: Failed to execute 'getFloatValue' on 'CSSPrimitiveValue': Failed to obtain a double value..
PASS getFloatValue(2, '-webkit-transform', CSSPrimitiveValue.CSS_HZ) threw exception InvalidAccessError: Failed to execute 'getFloatValue' on 'CSSPrimitiveValue': Failed to obtain a double value..
PASS getFloatValue(2, '-webkit-transform', CSSPrimitiveValue.CSS_RGBCOLOR) threw exception InvalidAccessError: Failed to execute 'getFloatValue' on 'CSSPrimitiveValue': Failed to obtain a double value..
PASS getFloatValue(1, '-webkit-animation-delay', CSSPrimitiveValue.CSS_PERCENTAGE) threw exception InvalidAccessError: Failed to execute 'getFloatValue' on 'CSSPrimitiveValue': Failed to obtain a double value..
PASS getFloatValue(1, '-webkit-animation-delay', CSSPrimitiveValue.CSS_PX) threw exception InvalidAccessError: Failed to execute 'getFloatValue' on 'CSSPrimitiveValue': Failed to obtain a double value..
PASS getFloatValue(1, '-webkit-animation-delay', CSSPrimitiveValue.CSS_DEG) threw exception InvalidAccessError: Failed to execute 'getFloatValue' on 'CSSPrimitiveValue': Failed to obtain a double value..
PASS getFloatValue(1, '-webkit-animation-delay', CSSPrimitiveValue.CSS_HZ) threw exception InvalidAccessError: Failed to execute 'getFloatValue' on 'CSSPrimitiveValue': Failed to obtain a double value..
PASS getFloatValue(1, '-webkit-animation-delay', CSSPrimitiveValue.CSS_RGBCOLOR) threw exception InvalidAccessError: Failed to execute 'getFloatValue' on 'CSSPrimitiveValue': Failed to obtain a double value..
PASS successfullyParsed is true

TEST COMPLETE
